% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_vis.R
\name{check_valid_ott}
\alias{check_valid_ott}
\title{Check if an OTT ID is valid in the Tree of Life}
\usage{
check_valid_ott(ott_id)
}
\arguments{
\item{ott_id}{A string representing the OTT ID to check}
}
\value{
A logical value indicating whether the OTT ID is valid
}
\description{
Check if an OTT ID is valid in the Tree of Life
}
