% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_vis.R
\name{chrom_vis}
\alias{chrom_vis}
\title{Visualize Chromosome Data}
\usage{
chrom_vis(chrom_data)
}
\arguments{
\item{chrom_data}{A data frame containing chromosome data with columns for chromosome names and their lengths.}
}
\value{
A horizontal bar chart visualizing chromosome lengths.
}
\description{
This function creates a horizontal bar chart to visualize chromosome lengths from a given dataset.
}
