% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_curation.R
\name{expand_chr_names}
\alias{expand_chr_names}
\title{Expand chromosome names in matches}
\usage{
expand_chr_names(matches)
}
\arguments{
\item{matches}{A data frame containing genome matches from `compare_genome_sigs`}
}
\description{
This function transforms genome signature comparison output to a long format
}
