% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_consensus.R
\name{genome_consensus}
\alias{genome_consensus}
\title{Physical consensus genomes}
\usage{
genome_consensus(genome_sorted)
}
\arguments{
\item{genome_sorted}{A list of sorted genomes, each containing chromosome names and lengths}
}
\value{
A consensus genome with averaged chromosome lengths
}
\description{
This function generates a consensus genome based on the physical chromosome lengths of multiple genomes.
}
