% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_curation.R
\name{map_chromosomes}
\alias{map_chromosomes}
\title{Map chromosomes in genome matches for homology visualization}
\usage{
map_chromosomes(matches)
}
\arguments{
\item{matches}{A data frame containing genome matches from `compare_genome_sigs`}
}
\description{
This function prepares genome comparison data for visualization by mapping chromosomes and their order patterns.
}
