% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_queries.R
\name{ncbi_genome_metadata}
\alias{ncbi_genome_metadata}
\title{Get NCBI genome metadata for a given taxonomic group}
\usage{
ncbi_genome_metadata(taxa, key)
}
\arguments{
\item{taxa}{A character string indicating the taxanomic group to query}

\item{key}{An character string for the NCBI API key}
}
\value{
A list containing genome assembly metadata
}
\description{
This function retrieves genome assembly metadata from NCBI for a specified taxonomic group including genome name, assembly accession, submitter organization, assembly type, and release date.
}
\examples{
kiwi <- ncbi_genome_metadata("Actinidia chinensis")
}
