% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_queries.R
\name{ncbi_genome_stats}
\alias{ncbi_genome_stats}
\title{Get NCBI genome statistics for a given taxonomic group}
\usage{
ncbi_genome_stats(taxa)
}
\arguments{
\item{taxa}{A character string indicating the taxanomic group to query}
}
\value{
A data.frame containing genome statistics including genome name and physical chromosome lengths
}
\description{
This function retrieves genome statistics from NCBI for a specified taxonomic group including genome name, total length, and other relevant information.
}
\examples{
kiwi <- ncbi_genome_stats("Actinidia chinensis")
}
