% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_queries.R
\name{ncbi_genome_stats_and_metadata}
\alias{ncbi_genome_stats_and_metadata}
\title{Assemble complete NCBI stats for one taxonomic group}
\usage{
ncbi_genome_stats_and_metadata(taxonomy, key)
}
\arguments{
\item{taxonomy}{A character string indicating the taxonomic group to query}

\item{key}{An character string for the NCBI API key}
}
\value{
a list containing genome statistics and metadata
}
\description{
This function combines genome statistics and metadata from NCBI for a specified taxonomic group into a single data frame.
}
