% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_vis.R
\name{plot_mapping}
\alias{plot_mapping}
\title{Plot alluvial diagram for chromosome mapping}
\usage{
plot_mapping(chromosome_maps, taxon_name = "")
}
\arguments{
\item{chromosome_maps}{A data frame containing chromosome mapping data with columns for genome, chromosome, HOM, and HOM_chr}

\item{taxon_name}{A string representing the name of the taxon for the plot title}
}
\value{
A ggplot object visualizing the chromosome mapping
}
\description{
Plot alluvial diagram for chromosome mapping
}
