% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_consensus.R
\name{sort_genomes}
\alias{sort_genomes}
\title{Sorts genomes}
\usage{
sort_genomes(genome_stats)
}
\arguments{
\item{genome_stats}{A data frame containing genome lengths returned by `ncbi_genome_stats`}
}
\value{
A list of genome names sorted by similarity
}
\description{
This function sorts genome versions based on similarity and returns a list of matching genomes.
}
