% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_vis.R
\name{species_donut}
\alias{species_donut}
\title{Visualize NCBI Assemblies by Species}
\usage{
species_donut(genome_metadata)
}
\arguments{
\item{genome_metadata}{A data frame containing genome metadata with a column for species.}
}
\value{
A donut chart visualizing the number of NCBI assemblies by species.
}
\description{
This function creates a donut chart to visualize the number of NCBI assemblies grouped by species.
}
