% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_vis.R
\name{tax_tree}
\alias{tax_tree}
\title{Visualize taxonomic trees relevant to genome data}
\usage{
tax_tree(genome_metadata)
}
\arguments{
\item{taxonomy}{A string representing the top level taxonomic group to visualize}
}
\value{
A tree chart visualizing the taxonomic relationships
}
\description{
This function creates a radial tree chart to visualize taxonomic relationships relevant to genome data.
}
